+++ kernel/src/mm/paged/address.rs
@@ -1,18 +1,18 @@
 #[repr(C)]
 #[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct PhysicalAddress(usize);
+pub struct PhysicalAddress(u64);
 
 impl PhysicalAddress {
     //
 }
 
-impl From<usize> for PhysicalAddress {
-    fn from(val: usize) -> Self {
+impl From<u64> for PhysicalAddress {
+    fn from(val: u64) -> Self {
         Self(val)
     }
 }
 
-impl From<PhysicalAddress> for usize {
+impl From<PhysicalAddress> for u64 {
     fn from(val: PhysicalAddress) -> Self {
         val.0
     }
@@ -20,10 +20,10 @@ impl From<PhysicalAddress> for usize {
 
 #[repr(C)]
 #[derive(Copy, Clone, Ord, PartialOrd, Eq, PartialEq)]
-pub struct VirtualAddress(usize);
+pub struct VirtualAddress(u64);
 
 impl VirtualAddress {
-    pub fn vpn(&self) -> (usize, usize, usize) {
+    pub fn vpn(&self) -> (u64, u64, u64) {
         (
             self.0 & 0x1FF000,
             self.0 & 0x3FE00000,
@@ -31,18 +31,18 @@ impl VirtualAddress {
         )
     }
 
-    pub fn offset(&self) -> usize {
+    pub fn offset(&self) -> u64 {
         self.0 & 0xFFF
     }
 }
 
-impl From<usize> for VirtualAddress {
-    fn from(val: usize) -> Self {
+impl From<u64> for VirtualAddress {
+    fn from(val: u64) -> Self {
         Self(val)
     }
 }
 
-impl From<VirtualAddress> for usize {
+impl From<VirtualAddress> for u64 {
     fn from(val: VirtualAddress) -> Self {
         val.0
     }
